+++ .github/scripts/test-linux.sh
@@ -27,18 +27,6 @@ ${PYBIN}/pip install ./dist/pyvirtualcam*cp${PYVER}*manylinux*.whl
 
 ${PYBIN}/pip install -r dev-requirements.txt
 
-mkdir tmp_for_test
-pushd tmp_for_test
-# NOTE: TESTING DISABLED!
-# The v4l2loopback kernel module cannot be installed as it depends
-# on v4l2 (videodev) kernel support.
-# videodev can either be built into the kernel directly or supported
-# as loadable module.
-# Azure (which is what GitHub Actions uses) only offers videodev
-# in linux-modules-extra-azure starting from Ubuntu 20.10.
-# See https://packages.ubuntu.com/search?suite=groovy&arch=any&mode=exactfilename&searchon=contents&keywords=videodev.ko
-# However, GitHub Actions only uses LTS releases of Ubuntu and they
-# are currently at 20.04 LTS. The next LTS will be 22.04 LTS
-# which will come out in 2022.
-#${PYBIN}/pytest -v -s /io/test
-popd
+# Device-dependent tests (test_camera.py, test_capture.py) require v4l2loopback
+# and are not run in CI. Mock/utility tests validate the Python API contract.
+${PYBIN}/pytest -v /io/test/test_backend_contract.py /io/test/test_util.py